requirements.txt
allure-pytest
asyncio
atlassian-python-api~=4.0.4
autopep8
beautifulsoup4~=4.13.4
boto3~=1.38.36
cryptocode~=0.1
delayed-assert
jproperties~=2.1.2
mutagen~=1.47.0
pip-chill
pre-commit
pydrive~=1.3.1
pyjwt~=2.10.1
pylint
pytest-xdist

pymysql~=1.1.1
pyperclip3
pysocks
pytest-asyncio
pytest-order
pytz~=2025.2
softest
sshtunnel~=0.4.0
webdriver-manager
selenium~=4.33.0
python-jenkins~=1.8.2
pyyaml~=6.0.2

pytest~=8.4.0
playwright~=1.52.0
paramiko~=3.5.1
requests~=2.32.4
cryptography~=45.0.4
logger~=1.4
allure-python-commons~=2.14.3